Move json types around
